+++ amuyu/lib/version_history.md
@@ -49,5 +49,7 @@
 ## Update - 7.1 -  06/07/2025
 
 - Añadido Actividades Diarias
+## Update - 7.2 -  06/07/2025
 
+- Añadido TimeLine